Move options to `options-storage.ts`
as suggested in https://github.com/bfred-it/webext-options-sync/issues/5
