+++ source/background.ts
@@ -1,46 +1,7 @@
-import OptionsSync, {Migration} from 'webext-options-sync';
 import {addContextMenu} from 'webext-domain-permission-toggle';
 import {addToFutureTabs} from 'webext-dynamic-content-scripts';
 import './libs/cache';
-
-export interface Options {
-	customCSS: string;
-	personalToken: string;
-	logging: boolean;
-	[featureName: string]: string | boolean;
-}
-
-const defaults: Options = {
-	customCSS: '',
-	personalToken: '',
-	logging: false
-};
-
-for (const feature of window.collectFeatures.keys()) {
-	defaults[`feature:${feature}`] = true;
-}
-
-new OptionsSync().define({
-	defaults,
-	migrations: [
-		// Drop this migration after June 20
-		options => {
-			if (typeof options.disabledFeatures !== 'string') {
-				return;
-			}
-
-			for (const feature of options.disabledFeatures.split(/s+/)) {
-				options[`feature:${feature}`] = false;
-			}
-		},
-
-		// Example to for renamed features:
-		// featureWasRenamed('fix-squash-and-merge-title', 'sync-pr-commit-title'), // Merged on April 22
-
-		// Removed features will be automatically removed from the options as well
-		OptionsSync.migrations.removeUnused
-	]
-});
+import './options-storage';
 
 browser.runtime.onMessage.addListener(async message => {
 	if (!message || message.action !== 'openAllInTabs') {
@@ -82,12 +43,3 @@ browser.runtime.onInstalled.addListener(async ({reason}) => {
 // GitHub Enterprise support
 addToFutureTabs();
 addContextMenu();
-
-// @ts-ignore because this is only needed sometimes
-function featureWasRenamed(from: string, to: string): Migration { // eslint-disable-line @typescript-eslint/no-unused-vars
-	return (options: AnyObject) => {
-		if (typeof options[`feature:${from}`] === 'boolean') {
-			options[`feature:${to}`] = options[`feature:${from}`];
-		}
-	};
-}
